refactor: introduce `vms/platformvm/warp/signertest`
