+++ vms/platformvm/warp/signer_test.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2019-2024, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package warp
+package warp_test
 
 import (
 	"testing"
@@ -11,16 +11,18 @@ import (
 	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/utils/constants"
 	"github.com/ava-labs/avalanchego/utils/crypto/bls"
+	"github.com/ava-labs/avalanchego/vms/platformvm/warp"
+	"github.com/ava-labs/avalanchego/vms/platformvm/warp/signertest"
 )
 
 func TestSigner(t *testing.T) {
-	for name, test := range SignerTests {
+	for name, test := range signertest.SignerTests {
 		t.Run(name, func(t *testing.T) {
 			sk, err := bls.NewSecretKey()
 			require.NoError(t, err)
 
 			chainID := ids.GenerateTestID()
-			s := NewSigner(sk, constants.UnitTestID, chainID)
+			s := warp.NewSigner(sk, constants.UnitTestID, chainID)
 
 			test(t, s, sk, constants.UnitTestID, chainID)
 		})
